requirements.txt
openai>=1.0.0
Pillow>=8.2.0
opencv-python>=4.5.0
numpy>=2.0.0